structField for go1.19
