+++ reflectx.go
@@ -195,7 +195,7 @@ func StructOf(fields []reflect.StructField) reflect.Type {
 	rt := totype(typ)
 	st := toStructType(rt)
 	for _, i := range anonymous {
-		st.fields[i].offsetEmbed |= 1
+		st.fields[i].setEmbedded()
 	}
 	for i, n := range underscore {
 		st.fields[i].name = n
