+++ structfield_go119.go
@@ -1,0 +1,23 @@
+//go:build go1.19
+// +build go1.19
+
+package reflectx
+
+// Struct field
+type structField struct {
+	name    name    // name is always non-empty
+	typ     *rtype  // type of field
+	_offset uintptr // byte offset of field
+}
+
+func (f *structField) offset() uintptr {
+	return f._offset
+}
+
+func (f *structField) embedded() bool {
+	return f.name.embedded()
+}
+
+func (f *structField) setEmbedded() {
+	f.name.setEmbedded()
+}
