+++ type.go
@@ -247,21 +247,6 @@ type sliceType struct {
 	elem *rtype // slice element type
 }
 
-// struct field
-type structField struct {
-	name        name    // name is always non-empty
-	typ         *rtype  // type of field
-	offsetEmbed uintptr // byte offset of field<<1 | isEmbedded
-}
-
-func (f *structField) offset() uintptr {
-	return f.offsetEmbed >> 1
-}
-
-func (f *structField) embedded() bool {
-	return f.offsetEmbed&1 != 0
-}
-
 // structType represents a struct type.
 type structType struct {
 	rtype
